componants, basic data viz, and react-bootstrap
