+++ package.json
@@ -13,11 +13,15 @@
   "dependencies": {
     "axios": "^0.24.0",
     "babel-polyfill": "^6.26.0",
+    "bootstrap": "^5.1.3",
     "css-loader": "^3.4.2",
+    "mini-css-extract-plugin": "^2.4.5",
     "react": "^16.13.1",
+    "react-bootstrap": "^2.0.3",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.6",
-    "redux-thunk": "^2.4.1"
+    "redux-thunk": "^2.4.1",
+    "victory": "^36.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
