+++ webpack.config.js
@@ -1,4 +1,5 @@
 const webpack = require('webpack'); // only add this if you don't have yet
+const path = require('path');
 
 // load client.env
 require('dotenv').config({ path: './client.env' }); 
@@ -13,7 +14,11 @@ module.exports = {
                 use: {
                     loader: "babel-loader"
                 }
-            } 
+            },
+            {
+                test: /\.(sass|less|css)$/,
+                loaders: ['style-loader', 'css-loader', 'less-loader']
+            }
         ]
     },
     entry: {
